+++ RELEASES.md
@@ -22,6 +22,11 @@ Version 1.8.0 (unreleased)
     added.
   * [`SmallFloat`][smf-1-8] and [`SmallComplex`][smc-1-8] can now be
     initialized with [`Special`][spe-1-8].
+  * New methods <code>[SmallFloat][smf-1-8]::[new][smf-n-1-8]</code>
+    and <code>[SmallComplex][smc-1-8]::[new][smc-n-1-8]</code> were
+    added.
+  * [`SmallFloat`][smf-1-8] and [`SmallComplex`][smc-1-8] now
+    implement [`Default`].
   * [`Integer`][int-1-8] now implements
     <code>[AsRef][`AsRef`]&lt;[\[][slice][limb_t][gmp-limb-1-2][\]][slice]&gt;</code>.
   * [`Float`][flo-1-8] now implements
@@ -47,7 +52,9 @@ Version 1.8.0 (unreleased)
 [oc-1-8]: https://docs.rs/rug/~1.8/rug/complex/struct.OrdComplex.html
 [of-1-8]: https://docs.rs/rug/~1.8/rug/float/struct.OrdFloat.html
 [smc-1-8]: https://docs.rs/rug/~1.8/rug/complex/struct.SmallComplex.html
+[smc-n-1-8]: https://docs.rs/rug/~1.8/rug/complex/struct.SmallComplex.html#method.new
 [smf-1-8]: https://docs.rs/rug/~1.8/rug/float/struct.SmallFloat.html
+[smf-n-1-8]: https://docs.rs/rug/~1.8/rug/float/struct.SmallFloat.html#method.new
 [smi-1-8]: https://docs.rs/rug/~1.8/rug/complex/struct.SmallInteger.html
 [smi-n-1-8]: https://docs.rs/rug/~1.8/rug/complex/struct.SmallInteger.html#method.new
 [smr-1-8]: https://docs.rs/rug/~1.8/rug/complex/struct.SmallRational.html
@@ -582,6 +589,7 @@ Version 0.6.0 (2017-08-09)
 [*Cargo.toml*]: https://doc.rust-lang.org/cargo/guide/dependencies.html
 [`AsMut`]: https://doc.rust-lang.org/nightly/core/convert/trait.AsMut.html
 [`AsRef`]: https://doc.rust-lang.org/nightly/core/convert/trait.AsRef.html
+[`Default`]: https://doc.rust-lang.org/nightly/core/default/trait.Default.html
 [`MaybeUninit`]: https://doc.rust-lang.org/nightly/core/mem/union.MaybeUninit.html
 [`Product`]: https://doc.rust-lang.org/nightly/core/iter/trait.Product.html
 [`RemAssign`]: https://doc.rust-lang.org/nightly/core/ops/trait.RemAssign.html